Répartir le code en plusieurs modules pour améliorer la lisibilité.
